Stacking Plates
Stacking plates problem where when your stack reaches a certain limit of plates, you create a new stack because you don't want the stack to grow large and for your plates to fall!

Implemented using a vector of pairs, where the first element is a stack and the second one is the element at which a stack is at.